+++ example/jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  preset: "jest-preset-angular",
+  snapshotSerializers: [
+    "jest-preset-angular/build/AngularNoNgAttributesSnapshotSerializer.js",
+    "jest-preset-angular/build/AngularSnapshotSerializer.js",
+    "jest-preset-angular/build/HTMLCommentSerializer.js"
+  ],
+  moduleNameMapper: {
+    "\\.(jpg|jpeg|png)$": "<rootDir>/__mocks__/image.js",
+    "^@lib/(.*)$": "<rootDir>/src/lib/$1"
+  },
+  setupFilesAfterEnv: [
+    "<rootDir>/src/setupJest.ts"
+  ]
+}